requirements.txt
pydantic==2.12.3
python-dotenv==1.2.1
PyYAML==6.0.3
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.63
langchain-openai==0.2.7
langchain-text-splitters==0.3.8
faiss-cpu==1.13.0
